chore: move modal component in centralized location to make it reusable (#634)
Currently, `modal` is only used by `hv-date-field`, but an upcoming
refactor will make it used by `hv-picker-field`. Move the code in a
shared location.
