+++ src/core/components/modal/overlay/styles.js
