+++ src/core/components/modal/styles.js
